Refactor: Moved the files around
