+++ src/types.ts
@@ -1,3 +1,3 @@
-import BetterArray from "./better_array.js";
+import BetterArray from "./Classes/better_array.js";
 
 export type ReadOnlyArray<T> = Omit<BetterArray<T>, 'push' | 'insert' | 'fill' | 'map' | 'pop' | 'reverse' | 'shift' | 'sort' | 'splice' | 'unshift' | 'toReadOnly'>;
